Initial commit of logitechmediaserver-bin-7.7.2-r2
This is currently identical to the ebuild in the main tree; the
intention is to remove that once this overlay is correctly established.

This migration has been prompted by the retirement of lavajoe and the
desire to avoid this ebuild in the main tree due to the large number of
bundled CPAN packages and libraries in the upstream release.