Corrected LICENSE and RESTRICT
License was inaccurately listed as GPL-2 (most, but not all, the
contents of the package were GPL). Corrected this to the special EULA
included with the package (this is included as a new license

logitechmediaserver-bin in the overlay), and also added RESTICT of
bindist and mirror to reflect the license's restrictions on
redistribution of the binary package.

See bug#450914